third_party/blink/web_tests/fast/scrolling/scrollbars/mouse-scrolling-on-div-scrollbar.html

<!DOCTYPE html>
<title>Tests mouse interactions on a non-custom composited div scrollbar.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/blink-coordinates-util.js"></script>
<script src="../../../resources/scrollbar-util.js"></script>
<script src="resources/mouse-scrolling-on-div-scrollbar.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-actions.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<style>
  .appearance {
    width: 100px;
    height: 100px;
    overflow: scroll;
    border: 1px solid black;
  }

  .standardLocation {
    position: absolute;
    top: 100px;
    left: 100px;
  }

  .space {
    height: 200px;
    width: 200px;
  }
</style>

<!-- Composited non-custom scroller -->
<div id="standard" class="appearance standardLocation" style="direction: ltr;">
  <div class="space"></div>
</div>

<script>
  if (window.internals)
    internals.settings.setScrollAnimatorEnabled(false);

  const scroller = document.getElementById("standard");

  assert_equals(scroller.clientHeight, scroller.clientWidth,
    "This test assumes that the height and width of the scroller are equivalent. If this changes please update SCROLL_AMOUNT to be X/Y specific");

  promise_test(async () => {
    await testArrows(scroller);
  }, "Test mouse click on non-custom composited div scrollbar arrows.");
  promise_test(async () => {
    await testTrackparts(scroller);
  }, "Test mouse click on non-custom composited div scrollbar empty trackparts.");
</script>
